fix:  Order 데이터 타입 수정, 마이바티스 경로 수정

+++ NareuGO_Backend/src/main/java/org/example/nareugobackend/api/controller/payment/request/PaymentConfirmRequestDto.java
@@ -2,6 +2,7 @@ package org.example.nareugobackend.api.controller.payment.request;
 
 import lombok.Getter;
 import lombok.Setter;
+import java.math.BigDecimal;
 
 // 프론트엔드에서 결제 승인 요청 시 백엔드로 보낼 데이터
 @Getter
@@ -10,5 +11,5 @@ public class PaymentConfirmRequestDto {
 
     private String paymentKey;
     private Long orderId;
-    private Integer amount;
+    private BigDecimal amount;
 }
